requirements/base.txt
django-axes==1.3.6
django-tastypie==0.12.0
stripe==1.9.3
MySQL-python==1.2.5
python-memcached==1.53
python-mimeparse==0.1.4
django-localflavor==1.0
